+++ src/components/DeployGroup.jsx
@@ -1,11 +1,20 @@
+import DeployCard from "./DeployCard";
+import { useState } from "react";
 
-import DeployListCard from "./DeployListCard";
-
+/**
+ * Display a 'Deploy group' (name and list of deployement)
+ * @param {*} props
+ * @returns
+ */
 function DeployGroup(props) {
   return (
-    <div className="border rounded m-2">
+    <div className="border rounded m-2" style={{ background: "#F8F9FA" }}>
       <h1>{props.groupname}</h1>
-      <DeployListCard deployList={props.deployList}></DeployListCard>
+      <div className="d-flex justify-content-around flex-wrap">
+        {props.deployList.map((deploy) => (
+          <DeployCard key={deploy.id} deploy={deploy}></DeployCard>
+        ))}
+      </div>
     </div>
   );
 }